Implement first interpreter (only assignment)
So this one had a lot of setup work. It is also kinda useless since
there are no function calls and so you cannot read the values
you set.

However the follwing example already works:
```
a = 3
b = a
```

For development purposes we currently print the environment after each
line in the REPL.

Very happy with the progress so far, build in functions are next.

+++ Sources/Moose/Environment/MooseObject.swift
@@ -13,9 +13,9 @@ protocol MooseObject: CustomStringConvertible {
 
 class IntegerObj: MooseObject {
     let type: MooseType = .Int
-    let value: Int?
+    let value: Int64?
 
-    init(value: Int?) {
+    init(value: Int64?) {
         self.value = value
     }
